Add files via upload
We put together this hands-on lab designed to get you familiar with all aspects of MongoDB from deploying a cluster and loading data, working with MongoDB's document data model, to using MongoDB Stitch, our serverless platform. 
First we will spin up a free database cluster using MongoDB Atlas - our database as a service. This is by far the easiest and fastest way to get you developing with MongoDB. 

Then we will install MongoDB Compass - this is our GUI for MongoDB which will allow you to visually explore and interact with your data with full CRUD (Create-Read-Update-Delete) functionality.

Using Compass we will create a database and load data to a collection - and it will also provide a nice context to demonstrate and discuss the power and flexibility of the document data model.